64-bit build target for the NPAPI plugin, and fixes for various size_t-related compiler warnings.
Review URL: http://codereview.chromium.org/552140

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1432 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 56d3b4ef7c7dba4533935ff4a83be687c5c175ff

+++ src/trusted/plugin/srpc/multimedia_socket.cc
@@ -25,7 +25,7 @@
 
 namespace nacl_srpc {
 
-int MultimediaSocket::kNaClMultimediaBridgeIdent;
+uintptr_t MultimediaSocket::kNaClMultimediaBridgeIdent;
 
 // NB: InitializeIdentifiers is not thread-safe.
 void MultimediaSocket::InitializeIdentifiers(
